Forward A8_M3 serial over ssh instead of 20000
Forward the a8-m3 serial link using ssh instead of two socat using port 20000.
In fact, only port 22 is allowed to be forwarded by the server iptables.
Also it now requires only one subprocess running instead of two.

It worked on port 20000 because the internal gateway switch was allowing it at
some point. But this was a bug and not intended. By 'chance' it was working.