Major modification of the class board_config
The singleton was not a real one and the content of _init was always executed.
This has been fixed and the configuration of the board is readed once in the execution
To enable test to work, a function clear_instance has been added. It allows to re-read the configuration if BoardConfig() is called latter
Finally the path the open node file is now realised thanks to .format