flash_timeout: Update timeout to 100seconds
When testing with big firmwares, the flash time can be really long.
Was 40s for an average firmware on samr21, I got 20 seconds with a big
firmware on m3.

The goal is to unstuck the node after some time not prevent flashing any files.
So I prefer putting a big value here.

I used 100 seconds here, as experiment handler currently has a 120s timeout.