integration_tests: wait node serial ready after power on
Using update/idle in check_debug made the code a bit faster which broke tests
on samr21 where the tty was not ready when flashing.

Now explicitely wait for the TTY to be ready, and still wait 1s for the
firmware to be started as the first test after could change.

This 'wait' is not part of the flash command, as flashing with a powered off
node or a still not ready node is normal to fail, you should wait or just retry.

But in tests I need reliability.