UPDATE CC2538 class now puts generated hex file in /tmp/
Changes the path of the temporary hex file. Now it is not in the same repository, but always in /tmp/ file.
This caused problem due to the teardown call function in node_firefly.py, which returned an error because of an impossibility to create the hex file in gateway_code/static/